add zsh support
"command cd" fails to run any "cd" in zsh, eventhough it should
according to POSIX. Instead, in zsh we need "builtin cd".

So use that. The name to use is chosen once - when cdd.sh is dotted.

Note that when (our override) "cd" is used, if it fails, then the
error message begins with "__cdd_builtin:cd: ..." (which is where we
call the builtin cd from), while without our "cd" override function
it's "cd: ...", however, it's not too bad.

In bash, both "command cd" and "cd" without our override print the
same error message, starting with "bash: cd: ..." .