Search (shell) pattern instead of only sub-string
Typically the search would be for a plain string anyway, so this is
unaffacted, but now we also support e.g. foo*bar, which will match
this sub-pattern, i.e. *foo*bar*, and this can be useful too.

The code change is tiny - unquote 2 places and remove -F with grep.
The rest is help/README updates, mainly related to this change.